Add files via upload
Brick Breaker on the theme of the Three Little Pigs. A project in my JavaScript journey. Since then, I've delved into learning Object-Oriented Programming (OOP) and identified areas for improvement to enhance gameplay, including the addition of levels and lives. The code also needs optimization with the incorporation of more and clearer functions, as well as well-defined classes. It was a truly great and enjoyable experience that allowed me to learn how to use Canvas for creating games or animations and to get started with JavaScript Objects.